+++ src/cde-configuration-service/cde-config.d.ts
