+++ src/cde-configuration-service/cde-runtime-config.ts
@@ -1,4 +1,4 @@
-import { CDEConfigurationService, CDEConfiguration, CDEDefaultConfiguration } from './';
+import { CDEConfigurationService, CDEConfiguration, CDEDefaultConfiguration } from '.';
 export class CDERuntimeConfig implements CDEConfigurationService {
 
     private defaultConfiguration : CDEConfiguration;
