+++ src/cde-configuration-service/model/cde-configuration.ts
