+++ src/cde-configuration-service/model/index.ts
