+++ src/model/disease-result.ts
