+++ src/model/disease-results.ts
