+++ src/model/intervention-result.ts
