+++ src/model/term-result.ts
