+++ src/model/term-results.ts
